backend-python/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.1
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.110.0
frozenlist==1.8.0
greenlet==3.3.2
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.14
langchain-community==0.0.38
langchain-core==1.2.25
langchain-groq==1.1.2
langchain-text-splitters==0.0.2
langgraph==1.1.4
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.24
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==1.26.4
orjson==3.11.8
ormsgpack==1.12.2
packaging==23.2
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.36.3
tenacity==8.5.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.28.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
